requirements.txt
requests==2.32.3
python-dotenv==1.0.1
tqdm==4.66.4